+++ .prettierrc
@@ -2,5 +2,9 @@
 	"useTabs": true,
 	"singleQuote": true,
 	"trailingComma": "none",
-	"printWidth": 120
+	"printWidth": 120,
+	"importOrderParserPlugins": ["typescript", "decorators-legacy"],
+	"importOrder": ["^@(.*)/(.*)$", "^[./]"],
+	"importOrderSeparation": true,
+	"importOrderSortSpecifiers": true
 }
