+++ package.json
@@ -21,14 +21,14 @@
 		"test:loadFixtures": "vitest initAll"
 	},
 	"dependencies": {
+		"@apollo/server": "4.5.0",
 		"@mikro-orm/core": "5.6.14",
 		"@mikro-orm/migrations": "5.6.14",
 		"@mikro-orm/postgresql": "5.6.14",
 		"@mikro-orm/sql-highlighter": "1.0.1",
-		"@apollo/server": "4.5.0",
+		"body-parser": "1.20.2",
 		"class-validator": "0.14.0",
 		"cors": "2.8.5",
-		"body-parser": "1.20.2",
 		"cross-env": "7.0.3",
 		"dotenv": "16.0.3",
 		"express": "4.18.2",
@@ -42,6 +42,7 @@
 	"devDependencies": {
 		"@faker-js/faker": "7.6.0",
 		"@mikro-orm/cli": "5.6.14",
+		"@trivago/prettier-plugin-sort-imports": "^4.1.1",
 		"@types/cors": "2.8.13",
 		"@types/express": "4.17.17",
 		"@types/node": "18.15.3",
@@ -49,19 +50,19 @@
 		"@types/uuid": "9.0.1",
 		"@typescript-eslint/eslint-plugin": "5.55.0",
 		"@typescript-eslint/parser": "5.55.0",
+		"@vitest/ui": "latest",
 		"eslint": "8.36.0",
 		"eslint-config-prettier": "8.7.0",
 		"eslint-plugin-markdown": "3.0.0",
 		"gen-env-types": "1.3.4",
+		"graphql-tag": "2.12.6",
 		"prettier": "2.8.4",
 		"supertest": "6.3.3",
 		"ts-node-dev": "2.0.0",
 		"tsconfig-paths": "4.1.2",
 		"typescript": "4.9.5",
-		"@vitest/ui": "latest",
-		"vitest": "0.29.2",
 		"vite": "4.1.4",
-		"graphql-tag": "2.12.6"
+		"vitest": "0.29.2"
 	},
 	"mikro-orm": {
 		"useTsNode": true,
