+++ src/entities/book.entity.ts
@@ -1,10 +1,11 @@
-import { Cascade, Collection, Entity, ManyToMany, ManyToOne, Property } from '@mikro-orm/core';
 import BookValidator from 'contracts/validators/Book.validator';
 import { Author } from 'entities/author.entity';
 import { Publisher } from 'entities/publisher.entity';
 import { Tag } from 'entities/tag.entity';
 import { Field, ObjectType } from 'type-graphql';
 import { Base } from 'utils/entities/base.entity';
+
+import { Cascade, Collection, Entity, ManyToMany, ManyToOne, Property } from '@mikro-orm/core';
 
 @ObjectType()
 @Entity()
