+++ src/entities/publisher.entity.ts
@@ -1,9 +1,10 @@
-import { Collection, Entity, Enum, OneToMany, Property } from '@mikro-orm/core';
 import { PublisherType } from 'contracts/enums/publisherType.enum';
 import { PublisherValidator } from 'contracts/validators/publisher.validator';
 import { Book } from 'entities/book.entity';
 import { Field, ObjectType } from 'type-graphql';
 import { Base } from 'utils/entities/base.entity';
+
+import { Collection, Entity, Enum, OneToMany, Property } from '@mikro-orm/core';
 
 @ObjectType()
 @Entity()
