+++ src/index.ts
@@ -2,6 +2,7 @@ import 'dotenv/config';
 import 'reflect-metadata';
 
 import Application from './application';
+
 (async () => {
 	const application = new Application();
 	await application.init();
