+++ src/utils/graphql-fields-to-relations/index.ts
@@ -1,6 +1,7 @@
 // this is https://github.com/bnussman/graphql-fields-to-relations
 // ported to graphql 16
 import { GraphQLResolveInfo } from 'graphql';
+
 import { graphqlFields } from './fields';
 
 export const fieldsToRelations = (
