+++ src/utils/interfaces/context.interface.ts
@@ -1,5 +1,6 @@
-import { Connection, EntityManager, IDatabaseDriver } from '@mikro-orm/core';
 import { Request, Response } from 'express';
+
+import { Connection, EntityManager, IDatabaseDriver } from '@mikro-orm/core';
 
 export interface MyContext {
 	req: Request;
